WIP file refactoring
`bio` will be a major namespace with many sub-modules.

I have tentatively sketched the following modules:

 * bio - types$ and (un)packers - no functions
 * bio.util - helpers, random things
 * bio.io - reading and writing bioinformatics formats
 * bio.db - read from databases
 * bio.plot - guess what?
 * bio.algo - bioinformatics algorithsm of every sort (maybe add sub-modules)

The `bio` module will be the most foundational that specifies the data
structures that all sub-modules use. `bio.util` will have general
helpers as well as some basic manipuations of biological data.

        bio
         |
         v
        util
         |
         +
    .---'|'-----.--------.
    |    |      |        |
    v    v      v        v
    db   io    plot     algo

        Dependency plot

I shall also need to distinguish between "modules" and "packages". A
module is one packet of code, with imports and exports. A package may
contain several modules and will typically be associated with a specific
github repo. The naming of packages will tend to have a single name for
core packages and two names (username + reponame) for community
libraries.

Packages have versions and all the `package.yaml` metadata, but modules
do not. So for the "bio" package, there is one version and one
description. I do eventually need to add documentation syntax to morloc
script, but that can wait. The "bio" module will store the conventions
that are common across the package.