Separate tree type into rooted and unrooted
The unrooted trees have fewer functions, though I may add more later.
Many of the rooted tree functions are not as clean in the unrooted form
since there is no special root node.

The main new function I added was a `midpoint` function that transforms
an unrooted tree to a rooted tree. I will need this in the flu case
study.